test(examples): move shared fixtures to a common folder so they can b… (#280)
* test(examples): move shared fixtures to a common folder so they can be reused, remove repeated code

* test(examples): move shared fixtures to a common folder so they can be reused, remove repeated code

* test(examples): move shared fixtures to a common folder so they can be reused, remove repeated code
